+++ igdectk/bootstrap/__init__.py
@@ -8,3 +8,5 @@ Bootstrap Django integration sub-package init.
 
 # Application Config for startup and more...
 default_app_config = __name__ + '.apps.IgdecTkBootstrap'
+
+PACKAGER = 'bootstrap'